% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh.R
\name{gh_query}
\alias{gh_query}
\alias{gh_lang}
\alias{gh_contrib}
\alias{gh_issue}
\alias{gh_comment}
\alias{gh_commit}
\title{GitHub repository queries}
\usage{
gh_lang(repo)

gh_contrib(repo)

gh_issue(repo)

gh_comment(repo)

gh_commit(repo)
}
\arguments{
\item{repo}{Name of the repository, in the form \code{"owner/repository"}.}
}
\value{
A \code{tibble} containing the tabulated API responses, or \code{NA} if the response
was empty or unsuccessful.
}
\description{
Get the languages, contributors, issues, comments, or commits of a single
repository from the GitHub API.
}
\details{
Queries use \code{\link[=ghq]{ghq()}}, a thin wrapper for \code{\link[gh:gh]{gh::gh()}} that ensures the query
always returns a value.
}
\examples{
gh_lang("joeroe/era")
gh_contrib("joeroe/era")
gh_issue("joeroe/era")
gh_comment("joeroe/era")
gh_commit("joeroe/era")
}
