% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh.R
\name{ghq}
\alias{ghq}
\title{Query the GitHub API}
\usage{
ghq(endpoint, ...)
}
\arguments{
\item{endpoint}{GitHub API endpoint.}

\item{...}{Other arguments to \code{\link[gh:gh]{gh::gh()}}, including name-value pairs of any
API parameters in \code{endpoint}.}
}
\description{
Thin wrapper for \code{\link[gh:gh]{gh::gh()}} that returns an NA if the response is empty or
if returns a 404 error (with a warning).
}
\details{
Queries where any value of \code{...} is \code{NA} will also return an \code{NA} row.
}
\examples{
ghq("GET /repos/{repo}/languages", repo = "joeroe/c14")
ghq("GET /repos/{repo}/languages", repo = "nonexistent/repo")
}
