% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh.R
\name{ght}
\alias{ght}
\title{Query the GitHub API (tabulated)}
\usage{
ght(endpoint, ...)
}
\arguments{
\item{endpoint}{GitHub API endpoint.}

\item{...}{Other arguments to \code{\link[gh:gh]{gh::gh()}}, including name-value pairs of any
API parameters in \code{endpoint}.}
}
\value{
If each response returns a single row, a \code{tibble} with one row per query
(producing a packed column if used in \code{mutate()}). Otherwise, a list of
\code{tibbles} for each query (producing a nested column if used in \code{mutate()}).
}
\description{
\code{\link[dplyr:mutate]{dplyr::mutate()}}-friendly wrapper for \code{\link[gh:gh]{gh::gh()}} that tabulates the response
and always returns a vector that matches the number of queries.
Vectorised over the named API parameters in \code{...}.
}
\details{
Empty responses return an \code{NA} row, with a warning if the request produced
a 404 error. Queries where any value of \code{...} is \code{NA} will also return an
\code{NA} row.
}
\examples{
ght("GET /repos/{owner}/{repo}/languages", owner = c("joeroe", "joeroe"), repo = c("era", "c14"))
}
