% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh.R
\name{ght}
\alias{ght}
\title{Query the GitHub API and return a table}
\usage{
ght(endpoint, ...)
}
\arguments{
\item{endpoint}{GitHub API endpoint.}

\item{...}{Other arguments to \code{\link[gh:gh]{gh::gh()}}, including name-value pairs of any
API parameters in \code{endpoint}.}
}
\value{
A data frame with one row per query, and one column per response vector.
}
\description{
Vectorised wrapper for \code{\link[gh:gh]{gh::gh()}} that:
\itemize{
\item Converts the response to a data frame
\item Silently skips queries where any value of \code{...} is \code{NA}, returning an \code{NA}
row.
\item Catches API error statuses (e.g. 404), returning an \code{NA} row and a printed
message instead of an error.
}
}
\examples{
ght("GET /repos/{owner}/{repo}/languages",
    owner = c("joeroe", "joeroe"), repo = c("era", "c14"))
}
